requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.24.0
Werkzeug==2.3.7
WTForms==3.2.1
